all indicators have been split with packages named after each indicator and a new calculate() method added
