+++ gctscript/examples/ta/atr.gct
@@ -1,13 +1,13 @@
 fmt := import("fmt")
 exch := import("exchange")
 t := import("times")
-ind := import("indicator/range")
+atr := import("indicator/atr")
 
 load := func() {
     start := t.add(t.now(), -t.hour*24*100)
     ohlcvData := exch.ohlcv("BTC Markets", "BTC-AUD", "-", "", start, t.now(), "1d")
 
-    ret := ind.atr(ohlcvData.candles, 14)
+    ret := atr.calculate(ohlcvData.candles, 14)
     fmt.println(ret)
 }
