+++ gctscript/examples/ta/rsi.gct
@@ -1,13 +1,13 @@
 fmt := import("fmt")
 exch := import("exchange")
 t := import("times")
-ind := import("indicator/rsi")
+rsi := import("indicator/rsi")
 
 load := func() {
    start := t.add(t.now(), -t.hour*24*30)
    ohlcvData := exch.ohlcv("BTC Markets", "BTC-AUD", "-", "", start, t.now(), "1d")
-   rsi := ind.rsi(ohlcvData.candles, 14)
-   fmt.println(rsi)
+   ret := rsi.calculate(ohlcvData.candles, 14)
+   fmt.println(ret)
 }
 
 load()
