+++ gctscript/wrappers/gct/exchange/exchange.go
@@ -210,6 +210,7 @@ func (e Exchange) WithdrawalCryptoFunds(exch string, request *withdraw.Request) 
 	return resp.Exchange.ID, nil
 }
 
+// OHLCV returns open high low close volume candles for requested exchange/pair/asset/start & end time
 func (e Exchange) OHLCV(exch string, pair currency.Pair, item asset.Item, start, end time.Time, interval time.Duration) (kline.Item, error) {
 	ex, err := e.GetExchange(exch)
 	if err != nil {
